+++ src/jsMain/kotlin/js_main.kt
@@ -54,10 +54,12 @@ suspend fun runStrudelDemo() {
     val playerOptions = KlangPlayer.Options(
         samples = samples,
         sampleRate = 48000,
-        cyclesPerSecond = 0.5,
+        cyclesPerSecond = 0.7,
     )
 
-    val pattern = StrudelPattern.Static.fromJson(tetrisJson)
+    val pattern = StrudelPattern.Static
+//        .fromJson(tetrisJson)
+        .fromJson(strangerThingsJson)
 
     console.log("Running strudel player")