+++ src/jvmMain/kotlin/Main.kt
@@ -68,7 +68,7 @@ private suspend fun helloStrudel() {
         val playerOptions = KlangPlayer.Options(
             samples = samples,
             sampleRate = 48_000,
-            cyclesPerSecond = 0.5,
+            cyclesPerSecond = 0.7,
         )
 
         val audio1 = strudelPlayer(